Add Deposit API

+++ src/main/java/com/flapkap/vendingmachine/dto/UserDTO.java
@@ -3,13 +3,11 @@ package com.flapkap.vendingmachine.dto;
 import com.flapkap.vendingmachine.model.Role;
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 @Builder
 public record UserDTO(
         String username,
         String password,
-        BigDecimal deposit,
+        Integer deposit,
         Role role,
         boolean enabled
 ) {
